Fix config_mask for shmem_team_split_strided in shmem_team_get_config.c
The team creation in shmem_team_get_config.c used `0` for `config_mask`,
but expected the team to be created with the user supplied `config`.
However, the spec says "A `config_mask` value of 0 indicates that the
team should be created with the default values for all configuration
parameters" and the default value for the number of contexts is 0.

Update the test to use `SHMEM_TEAM_NUM_CONTEXTS` for the mask and add
another test to validate that a mask of 0 defaults to 0 contexts.